link DB to dataprocessing tidyRules

+++ DataProcessing/Models/Output.cs
@@ -62,7 +62,7 @@ namespace DataProcessing.Models
     {
         public string? name { get; set; }
         public string? description { get; set; }
-        public int? weight { get; set; }
+        public double? weight { get; set; }
         public List<Configuration>? configurations { get; set; }
     }
 }
